feat: Move Color enum out of Pawn class
- Move Color enum to reuse it.
- So, Pawn.BLACK_COLOR and Pawn.WHITE_COLOR are useless.

- Color enum을 재사용하기 위해 Pawn 객체 밖으로 옮겼습니다.
- 그래서 Pawn.BLACK_COLOR와 Pawn.WHITE_COLOR가 필요 없을 것입니다.
